+++ GiEnJul.Test/ControllerTests/EventControllerTest.cs
@@ -1,5 +1,5 @@
 ﻿using GiEnJul.Controllers;
-using GiEnJul.Features;
+using GiEnJul.Repositories;
 using Moq;
 using System;
 using System.Threading.Tasks;
