+++ GiEnJul.Test/ControllerTests/GiverControllerTest.cs
@@ -1,7 +1,7 @@
 ﻿using GiEnJul.Clients;
 using GiEnJul.Controllers;
 using GiEnJul.Dtos;
-using GiEnJul.Features;
+using GiEnJul.Repositories;
 using Microsoft.AspNetCore.Mvc;
 using Moq;
 using System;
