+++ GiEnJul/Controllers/EventController.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using GiEnJul.Features;
+using GiEnJul.Repositories;
 using Microsoft.AspNetCore.Mvc;
 using Serilog;
 using System.Threading.Tasks;
