+++ GiEnJul/Controllers/GiverController.cs
@@ -1,8 +1,8 @@
 ﻿using AutoMapper;
 using GiEnJul.Clients;
 using GiEnJul.Dtos;
-using GiEnJul.Features;
 using GiEnJul.Models;
+using GiEnJul.Repositories;
 using Microsoft.AspNetCore.Mvc;
 using Serilog;
 using System.Threading.Tasks;
