+++ GiEnJul/Controllers/VerifyConnectionController.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 using GiEnJul.Exceptions;
-using GiEnJul.Features;
 using GiEnJul.Helpers;
+using GiEnJul.Repositories;
 using Microsoft.AspNetCore.Mvc;
 using Serilog;
 using System;
