+++ GiEnJul/Repositories/ConnectionRepository.cs
@@ -4,11 +4,11 @@ using GiEnJul.Infrastructure;
 using Serilog;
 using System.Threading.Tasks;
 
-namespace GiEnJul.Features
+namespace GiEnJul.Repositories
 {
     public interface IConnectionRepository
     {
-        Task<(string,string)> InsertOrReplaceAsync(Models.Giver giver, Models.Recipient recipient);
+        Task<(string, string)> InsertOrReplaceAsync(Models.Giver giver, Models.Recipient recipient);
         Task DeleteConnectionAsync(string partitionKey, string rowKey);
     }
     public class ConnectionRepository : GenericRepository<Connection>, IConnectionRepository
@@ -16,7 +16,7 @@ namespace GiEnJul.Features
         public ConnectionRepository(ISettings settings, IMapper mapper, ILogger log, string tableName = "Connection") : base(settings, tableName, mapper, log)
         { }
 
-        public async Task<(string,string)> InsertOrReplaceAsync(Models.Giver giver, Models.Recipient recipient)
+        public async Task<(string, string)> InsertOrReplaceAsync(Models.Giver giver, Models.Recipient recipient)
         {
             var connection = new Connection(_mapper.Map<Giver>(giver), _mapper.Map<Recipient>(recipient));
             await InsertOrReplaceAsync(connection);
