+++ GiEnJul/Repositories/GenericRepository.cs
@@ -7,7 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace GiEnJul.Features
+namespace GiEnJul.Repositories
 {
     public class GenericRepository<T> where T : EntityBase, new()
     {
