+++ GiEnJul/Repositories/RecipientRepository.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.Azure.Cosmos.Table;
 using System.Collections.Generic;
 
-namespace GiEnJul.Features
+namespace GiEnJul.Repositories
 {
     public interface IRecipientRepository
     {
@@ -31,12 +31,14 @@ namespace GiEnJul.Features
             var deleted = await DeleteAsync(_mapper.Map<Entities.Recipient>(model));
             return _mapper.Map<Models.Recipient>(deleted);
         }
-        public async Task<List<Models.Recipient>> GetUnmatchedRecipientsAsync( string location, string currentEvent )
+        public async Task<List<Models.Recipient>> GetUnmatchedRecipientsAsync(string location, string currentEvent)
         {
-            var query = new TableQuery<Entities.Recipient>() { FilterString = 
+            var query = new TableQuery<Entities.Recipient>()
+            {
+                FilterString =
                 TableQuery.CombineFilters(
-                    TableQuery.GenerateFilterCondition("PartitionKey", QueryComparisons.Equal, $"{currentEvent}_{location}"), 
-                    TableOperators.And, 
+                    TableQuery.GenerateFilterCondition("PartitionKey", QueryComparisons.Equal, $"{currentEvent}_{location}"),
+                    TableOperators.And,
                     TableQuery.GenerateFilterConditionForBool("IsMatched", QueryComparisons.Equal, false))
             };
